backend/requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
sqlalchemy[asyncio]
asyncpg
greenlet
PyJWT
stripe
aiofiles
httpx
